requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
aiofiles==23.2.1
boto3==1.34.108
python-docx==1.1.0
openai==0.27.10
redis>=4.5.0
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.1
aioresponses==0.7.6
moto[all]==4.2.13
python-dotenv==1.0.0
starlette==0.27.0
slowapi==0.1.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4